--- a/Makefile
+++ b/Makefile
@@ -9,3 +9,12 @@ init:
 .PHONY: test
 test:
 	go test -v ./...
+
+.PHONY: cover
+cover:
+	go test -coverpkg=. -covermode=atomic -coverprofile=coverage.txt
+
+.PHONY: view-cover
+view-cover: cover
+	go tool cover -html coverage.txt
+