--- a/callstack.go
+++ b/callstack.go
@@ -11,28 +11,66 @@ import (
 )
 
 // CallStack represents call stack.
-type CallStack []Frame
+type CallStack interface {
+	HeadFrame() Frame
+	Frames() []Frame
+}
+
+type callStack struct {
+	pcs []uintptr
+}
+
+func (cs callStack) HeadFrame() Frame {
+	if len(cs.pcs) == 0 {
+		return emptyFrame
+	}
+
+	rfs := runtime.CallersFrames(cs.pcs[:1])
+	f, _ := rfs.Next()
+	return frame{f.File, f.Line, f.Function}
+}
+
+func (cs callStack) Frames() []Frame {
+	if len(cs.pcs) == 0 {
+		return nil
+	}
+
+	rfs := runtime.CallersFrames(cs.pcs)
+
+	fs := make([]Frame, 0, len(cs.pcs))
+	for {
+		f, more := rfs.Next()
+
+		fs = append(fs, frame{f.File, f.Line, f.Function})
+
+		if !more {
+			break
+		}
+	}
+	return fs
+}
 
 // Format implements fmt.Formatter.
-func (cs CallStack) Format(s fmt.State, verb rune) {
+func (cs callStack) Format(s fmt.State, verb rune) {
 	switch verb {
 	case 'v':
 		switch {
 		case s.Flag('+'):
-			for _, pc := range cs {
-				fmt.Fprintf(s, "%+v\n", pc)
+			for _, f := range cs.Frames() {
+				fmt.Fprintf(s, "%+v\n", f)
 			}
 		case s.Flag('#'):
-			fmt.Fprintf(s, "%#v", []Frame(cs))
+			fmt.Fprintf(s, "%#v", cs.Frames())
 		default:
-			l := len(cs)
+			fs := cs.Frames()
+			l := len(fs)
 			if l == 0 {
 				return
 			}
-			for _, pc := range cs[:l-1] {
-				fmt.Fprintf(s, "%s: ", pc.Func())
+			for _, f := range fs[:l-1] {
+				fmt.Fprintf(s, "%s: ", f.Func())
 			}
-			fmt.Fprintf(s, "%v", cs[l-1].Func())
+			fmt.Fprintf(s, "%v", fs[l-1].Func())
 		}
 	case 's':
 		fmt.Fprintf(s, "%v", cs)
@@ -47,20 +85,7 @@ func Callers(skip int) CallStack {
 		return nil
 	}
 
-	fs := runtime.CallersFrames(pcs[:n])
-
-	cs := make(CallStack, 0, n)
-	for {
-		f, more := fs.Next()
-
-		cs = append(cs, frame{f.File, f.Line, f.Function})
-
-		if !more {
-			break
-		}
-	}
-
-	return cs
+	return callStack{pcs[:n]}
 }
 
 // CallStackFromPkgErrors creates CallStack from errors.StackTrace.
@@ -70,19 +95,7 @@ func CallStackFromPkgErrors(st errors.StackTrace) CallStack {
 		pcs[i] = uintptr(v)
 	}
 
-	fs := runtime.CallersFrames(pcs)
-
-	cs := make(CallStack, 0, len(pcs))
-	for {
-		f, more := fs.Next()
-
-		cs = append(cs, frame{f.File, f.Line, f.Function})
-
-		if !more {
-			break
-		}
-	}
-	return cs
+	return callStack{[]uintptr(pcs)}
 }
 
 type Frame interface {
@@ -93,7 +106,8 @@ type Frame interface {
 	Pkg() string
 }
 
-// PC represents program counter.
+var emptyFrame = frame{"???", 0, "???"}
+
 type frame struct {
 	file     string
 	line     int