--- a/callstack_test.go
+++ b/callstack_test.go
@@ -15,19 +15,19 @@ func X() failure.CallStack {
 }
 
 func TestCallers(t *testing.T) {
-	cs := X()
+	fs := X().Frames()
 
-	assert.Contains(t, cs[0].Path(), "github.com/morikuni/failure/callstack_test.go")
-	assert.Contains(t, cs[0].File(), "callstack_test.go")
-	assert.Equal(t, cs[0].Func(), "X")
-	assert.Equal(t, cs[0].Line(), 14)
-	assert.Equal(t, cs[0].Pkg(), "failure_test")
+	assert.Contains(t, fs[0].Path(), "github.com/morikuni/failure/callstack_test.go")
+	assert.Contains(t, fs[0].File(), "callstack_test.go")
+	assert.Equal(t, fs[0].Func(), "X")
+	assert.Equal(t, fs[0].Line(), 14)
+	assert.Equal(t, fs[0].Pkg(), "failure_test")
 
-	assert.Contains(t, cs[1].Path(), "github.com/morikuni/failure/callstack_test.go")
-	assert.Contains(t, cs[0].File(), "callstack_test.go")
-	assert.Equal(t, cs[1].Func(), "TestCallers")
-	assert.Equal(t, cs[1].Line(), 18)
-	assert.Equal(t, cs[1].Pkg(), "failure_test")
+	assert.Contains(t, fs[1].Path(), "github.com/morikuni/failure/callstack_test.go")
+	assert.Contains(t, fs[1].File(), "callstack_test.go")
+	assert.Equal(t, fs[1].Func(), "TestCallers")
+	assert.Equal(t, fs[1].Line(), 18)
+	assert.Equal(t, fs[1].Pkg(), "failure_test")
 }
 
 func Y() error {
@@ -43,58 +43,59 @@ func TestCallStackFromPkgErrors(t *testing.T) {
 	st, ok := err.(StackTracer)
 	require.True(t, ok)
 
-	cs := failure.CallStackFromPkgErrors(st.StackTrace())
+	fs := failure.CallStackFromPkgErrors(st.StackTrace()).Frames()
 
-	assert.Contains(t, cs[0].Path(), "github.com/morikuni/failure/callstack_test.go")
-	assert.Contains(t, cs[0].File(), "callstack_test.go")
-	assert.Equal(t, cs[0].Func(), "Y")
-	assert.Equal(t, cs[0].Line(), 34)
-	assert.Equal(t, cs[0].Pkg(), "failure_test")
+	assert.Contains(t, fs[0].Path(), "github.com/morikuni/failure/callstack_test.go")
+	assert.Contains(t, fs[0].File(), "callstack_test.go")
+	assert.Equal(t, fs[0].Func(), "Y")
+	assert.Equal(t, fs[0].Line(), 34)
+	assert.Equal(t, fs[0].Pkg(), "failure_test")
 
-	assert.Contains(t, cs[1].Path(), "github.com/morikuni/failure/callstack_test.go")
-	assert.Contains(t, cs[0].File(), "callstack_test.go")
-	assert.Equal(t, cs[1].Func(), "TestCallStackFromPkgErrors")
-	assert.Equal(t, cs[1].Line(), 42)
-	assert.Equal(t, cs[1].Pkg(), "failure_test")
+	assert.Contains(t, fs[1].Path(), "github.com/morikuni/failure/callstack_test.go")
+	assert.Contains(t, fs[1].File(), "callstack_test.go")
+	assert.Equal(t, fs[1].Func(), "TestCallStackFromPkgErrors")
+	assert.Equal(t, fs[1].Line(), 42)
+	assert.Equal(t, fs[1].Pkg(), "failure_test")
 }
 
 func TestFormat(t *testing.T) {
-	cs := X()[:2]
+	cs := X()
 
-	assert.Equal(t,
-		`X: TestFormat`,
+	assert.Regexp(t,
+		`X: TestFormat: .*`,
 		fmt.Sprintf("%v", cs),
 	)
-	assert.Equal(t,
-		`X: TestFormat`,
+	assert.Regexp(t,
+		`X: TestFormat: .*`,
 		fmt.Sprintf("%s", cs),
 	)
 	assert.Regexp(t,
-		`\[\]failure.Frame{/.+/github.com/morikuni/failure/callstack_test.go:14, /.+/github.com/morikuni/failure/callstack_test.go:62}`,
+		`\[\]failure.Frame{/.+/github.com/morikuni/failure/callstack_test.go:14, /.+/github.com/morikuni/failure/callstack_test.go:62, .*}`,
 		fmt.Sprintf("%#v", cs),
 	)
 	assert.Regexp(t,
 		`\[X\] /.+/github.com/morikuni/failure/callstack_test.go:14
-\[TestFormat\] /.+/github.com/morikuni/failure/callstack_test.go:62`,
+\[TestFormat\] /.+/github.com/morikuni/failure/callstack_test.go:62
+\[.*`,
 		fmt.Sprintf("%+v", cs),
 	)
 
-	pc := cs[0]
+	f := cs.Frames()[0]
 
 	assert.Regexp(t,
 		`/.+/github.com/morikuni/failure/callstack_test.go:14`,
-		fmt.Sprintf("%v", pc),
+		fmt.Sprintf("%v", f),
 	)
 	assert.Regexp(t,
 		`/.+/github.com/morikuni/failure/callstack_test.go:14`,
-		fmt.Sprintf("%s", pc),
+		fmt.Sprintf("%s", f),
 	)
 	assert.Regexp(t,
 		`/.+/github.com/morikuni/failure/callstack_test.go:14`,
-		fmt.Sprintf("%#v", pc),
+		fmt.Sprintf("%#v", f),
 	)
 	assert.Regexp(t,
 		`\[X\] /.+/github.com/morikuni/failure/callstack_test.go:14`,
-		fmt.Sprintf("%+v", pc),
+		fmt.Sprintf("%+v", f),
 	)
 }