Every Frame has a file path, function name, and line number.
Every CallStack has a list of Frames.
Failure chain contains some of CallStacks.

It allocates memory of `<size of a frame> * <size of call stack> * <length of failure chain>`, even if almost Frames are not used.

This PR reduces memory usage by lazy evaluation.

## benchmark

### on master

```
$ go test -bench . -benchmem -count 10
goos: darwin
goarch: amd64
pkg: github.com/morikuni/failure
BenchmarkFailure-4        200000              7070 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        200000              5260 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5215 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5187 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5251 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5191 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        200000              5213 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5302 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5277 ns/op            1968 B/op         21 allocs/op
BenchmarkFailure-4        300000              5465 ns/op            1968 B/op         21 allocs/op
PASS
ok      github.com/morikuni/failure     15.277s
```

### on this branch

```
$ go test -bench . -benchmem -count 10
goos: darwin
goarch: amd64
pkg: github.com/morikuni/failure
BenchmarkFailure-4        500000              2328 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1784 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1743 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1727 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1752 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1785 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1710 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1789 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1756 ns/op             864 B/op          6 allocs/op
BenchmarkFailure-4       1000000              1769 ns/op             864 B/op          6 allocs/op
PASS
ok      github.com/morikuni/failure     17.241s
```

3x faster and half memory usage.